phase-4-kubernetes/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend-phase3"
version = "0.2.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.1",
    "sqlmodel==0.0.32",
    "psycopg2-binary==2.9.11",
    "alembic==1.14.0",
    "pyjwt==2.11.0",
    "python-jose[cryptography]==3.3.0",
    "argon2-cffi==23.1.0",
    "pydantic[email]==2.9.2",
    "pydantic-settings==2.6.1",
    "python-multipart==0.0.22",
    "httpx==0.28.1",
    "openai-agents==0.8.1",
    "litellm>=1.0.0",
    "openai==2.9.0",
]
